Big changes + "company" logo!
I was mistakingly trying to commit a local version that wasn't linked to the github website but I think I finally got it. The layout turned out to look a bit "cheap" but this is my best at the moment. Hope to change is again but committing just for now.